requirements.txt
adium-theme-ubuntu==0.3.4
backports.functools-lru-cache==1.4
backports.weakref==1.0.post1
bleach==1.5.0
cycler==0.10.0
decorator==4.2.1
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
html5lib==0.9999999
Markdown==2.6.11
matplotlib==2.1.1
mock==2.0.0
networkx==2.0
numpy==1.14.0
pbr==3.1.1
Pillow==5.0.0
pkg-resources==0.0.0
protobuf==3.5.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
scikit-image==0.13.1
scipy==1.0.0
six==1.11.0
subprocess32==3.2.7
tensorflow-gpu==1.4.1
tensorflow-tensorboard==0.4.0rc3
unity-lens-photos==1.0
vboxapi==1.0
virtualenv==15.0.1
Werkzeug==0.14.1
